added new file parse.go to handle low level parsing of commands
